backend-submission/requirements.txt
fastapi
uvicorn
langchain
langchain_core
langchain_huggingface
pydantic
python-dotenv
redis
transformers
torch
huggingface_hub